backend/requirements.txt
Flask
flask-cors
langchain
langchain-community
faiss-cpu
openai
numpy
langchain-openai
joblib
scikit-learn